# Speed up lm test