main (compute hessian, don't use it, `newton-cg`):
```
[75.00%] ··· tests.Tests.time_inverse_lagrange_multiplier_test                                                                                                                                              ok
[75.00%] ··· ====== ============== ================ ============== =======================
             --                                      data                                 
             ------ ----------------------------------------------------------------------
               n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
             ====== ============== ================ ============== =======================
              1000    14.1±0.1ms       110±20ms          n/a             3.58±0.08ms      
             ====== ============== ================ ============== =======================

[100.00%] ··· tests.Tests.time_lagrange_multiplier_test                                                                                                                                                      ok
[100.00%] ··· ====== ============== ================ ============== =======================
              --                                      data                                 
              ------ ----------------------------------------------------------------------
                n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
              ====== ============== ================ ============== =======================
               1000     152±2μs         126±50ms       10.3±0.3ms          7.14±0.3ms      
              ====== ============== ================ ============== =======================
```

`37f2ab4` (compute hessian, use it, `newton-cg`):
```
[75.00%] ··· tests.Tests.time_inverse_lagrange_multiplier_test                                                                                                                                              ok
[75.00%] ··· ====== ============== ================ ============== =======================
             --                                      data                                 
             ------ ----------------------------------------------------------------------
               n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
             ====== ============== ================ ============== =======================
              1000    14.1±0.2ms       115±30ms          n/a              3.71±0.3ms      
             ====== ============== ================ ============== =======================

[100.00%] ··· tests.Tests.time_lagrange_multiplier_test                                                                                                                                                      ok
[100.00%] ··· ====== ============== ================ ============== =======================
              --                                      data                                 
              ------ ----------------------------------------------------------------------
                n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
              ====== ============== ================ ============== =======================
               1000    196±100μs        102±30ms       4.82±0.1ms          4.93±0.2ms      
              ====== ============== ================ ============== =======================
```

`5cfb2d1` (don't compute hessian, `newton-cg`):
```
[75.00%] ··· tests.Tests.time_inverse_lagrange_multiplier_test                                                                                                                                              ok
[75.00%] ··· ====== ============== ================ ============== =======================
             --                                      data                                 
             ------ ----------------------------------------------------------------------
               n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
             ====== ============== ================ ============== =======================
              1000   14.0±0.08ms      70.7±20ms          n/a             2.20±0.04ms      
             ====== ============== ================ ============== =======================

[100.00%] ··· tests.Tests.time_lagrange_multiplier_test                                                                                                                                                      ok
[100.00%] ··· ====== ============== ================ ============== =======================
              --                                      data                                 
              ------ ----------------------------------------------------------------------
                n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
              ====== ============== ================ ============== =======================
               1000     151±2μs        28.7±10ms      4.87±0.05ms         3.73±0.06ms      
              ====== ============== ================ ============== =======================
```

`bc7f67f`(don't compute hessian, `bfgs`):
```
[75.00%] ··· tests.Tests.time_inverse_lagrange_multiplier_test                                                                                                                                              ok
[75.00%] ··· ====== ============== ================ ============== =======================
             --                                      data                                 
             ------ ----------------------------------------------------------------------
               n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
             ====== ============== ================ ============== =======================
              1000    13.7±0.1ms      74.5±10ms          n/a              19.5±0.2ms      
             ====== ============== ================ ============== =======================

[100.00%] ··· tests.Tests.time_lagrange_multiplier_test                                                                                                                                                      ok
[100.00%] ··· ====== ============== ================ ============== =======================
              --                                      data                                 
              ------ ----------------------------------------------------------------------
                n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
              ====== ============== ================ ============== =======================
               1000     149±3μs        82.3±40ms      3.49±0.03ms         4.10±0.06ms      
              ====== ============== ================ ============== =======================
```